api: random: replace ssize_t with int32_t
ssize_t is a POSIX type. API definition must be pure C (C99).
Signed 32 bits can hold larger values than size_t (64k) and
ssize_t(32k) in minimum.

Reviewed-and-tested-by: Bill Fischofer <bill.fischofer@linaro.org>

+++ example/ipsec/odp_ipsec_cache.c
@@ -96,9 +96,9 @@ int create_ipsec_cache_entry(sa_db_entry_t *cipher_sa,
 
 	/* Generate an IV */
 	if (params.iv.length) {
-		ssize_t size = params.iv.length;
+		int32_t size = params.iv.length;
 
-		ssize_t ret = odp_random_data(params.iv.data, size, 1);
+		int32_t ret = odp_random_data(params.iv.data, size, 1);
 		if (ret != size)
 			return -1;
 	}
